Initial import of the memory layout tool from Pyromaniac.
The memory layout diagram tool is intended to allow different forms
of memory layout to be generated to allow some simple diagrams to
be produced to describe systems. The code is currently only usable
programatically, but it's not that complicated to work with. It
will eventually have a file format that you can use to make it easier
to work with.