Clean up the BBC memory map example. Allow removable labels.
Labels can now be removed. This has been applied to the BBC memory map
together with the elements which are on the right side of the diagram
in the source document. The size of the memory map has been reduced by
making the region's heights smaller than they were previously.