Add support for double lined regions, in SVG.
It is now possible to have a double lined region at the boundary.
This is only effective in the SVG output, as the Dot output doesn't
support this, but that's where it's most useful anyhow. It's also
possible to have no line there, but this will also mean that the
next region needs to be modified as well.