Add a cut-out style for the discontinuity in SVG.
The style of the discontinuity can be configured to allow them to be
given styles which match the effect that you want. Both the cut-out
and the zig-zag styles look pretty good to me. They could probably
also have a dashed block as well, but that hasn't been done yet.