Improve discontinuities; add elite-bbc example; font colouring.
Discontinuities are now better tested with an example being able to
be generated.

The Elite BBC cassette memory map now is now reproduced as a diagram,
which exercised the colouring and some of the additional features of
the rendering in dot and SVG.

And the font colouring is now supported.