Move the memory layout classes into a separate package.
Rather than being at the root the tools are now in a package.
This may make it easier to manage as a global package in the
future, but also means that it's simpler to work with - the
example code lives at the top level, and the library is a
separate thing.