Add a tool to process a file format describing the memory layout.
The memory layout is now able to be defined in a YAML file. It's not
a very well defined YAML structure, but an example has been created
for the BBC memory map, and it's not too terrible to work with.
Because there are some special regions in the memory map it's necessary
to have some new fatures like omitted address labels, but it generally
looks pretty sane.