Added ARM MCU example, and fixes for positioning.
The ARM MCU example isn't too complicated, but includes some features
on the left and the right, and a region that has marker ticks in it
which wasn't supported. Those junction marks are now supported, so
that we can inlcude logical divisions in that way, if that's what you
want.