Add line cap specifications for parts of the SVG; better widths.
Some parts of the SVG weren't covering the whole area, due to not
having line caps. I thought this was a reason that we were getting
some odd aliasing effects at the edges of the junctions, but it
wasn't and that indicated that we weren't even drawing the lines
to the end. David Thomas reported the odd effects in the PNGs, but
I still cannot see why they're there.

Some of the lines weren't drawn the right way - the dashed lines
were drawn top to bottom on one side and bottom to top on the other,
which produced a misaligned pattern. The widths were missing on
some elements too.