Move filled regions to be rendered before all other regions.
The filled regions would be drawn alongside the strokes which outline
the memory regions. This meant that a filled region might overdraw
the outline that had been drawn for the prior region. This mostly
affected the dotted and dashed junction boundaries, making the line
width appear half that of the vertical edges of the memory regions.

This change moves the filled elements to the start of the group,
so that they never overdraw the other borders, which occur later.